# TraceQL: support mixed-type attribute querying (int/float)