@joe-elliott 2025-01-02
I apologize for taking so long to get to this. Your analysis is correct! We do generate predicates per column and, since we store integers and floats independently we only scan one of the columns. Given how small int and float columns tend to be (compared to string columns) I think the performance hit of doing this is likely acceptable in exchange for the nicer behavior.

What is the behavior in this case? I'm pretty sure this will work b/c the engine will request all values for the two attributes and do the work itself. I believe the [engine layer](https://github.com/grafana/tempo/blob/main/pkg/traceql/ast_execute.go#L328) will compare ints and floats correctly but I'm not 100% sure.
  ```
  { span.intAttr > span.floatAttr }
  ```

Tests should also be [added here](https://github.com/grafana/tempo/blob/main/tempodb/tempodb_search_test.go#L45) for the new behavior. These tests build a block and then search for a known trace using a large range of traceql queries. If you add tests here and they pass it means that your changes work from the parquet file all the way up through the engine.

This will also break the "allConditions" optimization if the user types any query with a number comparison:

https://github.com/grafana/tempo/pull/4391/files#diff-a201423ab0b50d4455a497bf1804b1a9f596394413c28b7702710f89237c49c1R2815-R2821

I would like preserve the allConditions behavior in this case b/c it's such a nice optimization and number queries are common. I'm not quite sure why the `len(valueIters) == 1` condition exists so we'd need to do some research into it.