@ndk 2025-01-12
> I apologize for taking so long to get to this. Your analysis is correct! We do generate predicates per column and, since we store integers and floats independently we only scan one of the columns. Given how small int and float columns tend to be (compared to string columns) I think the performance hit of doing this is likely acceptable in exchange for the nicer behavior.

Thank you for confirming the approach and pointing out the `allConditions` optimization.  Right now, the fix scans both integer and float columns for attributes that might be either type. I’ve also adjusted how float comparisons work for integer fields, taking into account the fraction part and the comparison operator.

> What is the behavior in this case? I'm pretty sure this will work b/c the engine will request all values for the two attributes and do the work itself. I believe the [engine layer](https://github.com/grafana/tempo/blob/main/pkg/traceql/ast_execute.go#L328) will compare ints and floats correctly but I'm not 100% sure.
> 
> ```
> { span.intAttr > span.floatAttr }
> ```

I verified that `{ span.intAttr > span.floatAttr }` behaves as expected. Wanna me to add a test to cover this case?

> Tests should also be [added here](https://github.com/grafana/tempo/blob/main/tempodb/tempodb_search_test.go#L45) for the new behavior. These tests build a block and then search for a known trace using a large range of traceql queries. If you add tests here and they pass it means that your changes work from the parquet file all the way up through the engine.

Done. Let me know if I missed something.

> This will also break the "allConditions" optimization if the user types any query with a number comparison:
> 
> https://github.com/grafana/tempo/pull/4391/files#diff-a201423ab0b50d4455a497bf1804b1a9f596394413c28b7702710f89237c49c1R2815-R2821
> 
> I would like preserve the allConditions behavior in this case b/c it's such a nice optimization and number queries are common. I'm not quite sure why the `len(valueIters) == 1` condition exists so we'd need to do some research into it.

Regarding the `allConditions` block, the optimization is lost because we generate two predicates (one for int, one for float) under the same attribute name, triggering a `LeftJoinIterator` instead of a `JoinIterator.` Possible workarounds I’m considering:

- Creating a variant of `JoinIterator` that uses logical `OR` rather than `AND.`
- Exploring `parquet.multiRowGroup`, `parquetquery.UnionIterator`, or `parquetquery.KeyValueGroupPredicate` to see if they can unify the int/float search without losing the optimization.
- ~~Refactoring single working with `ColumnChunk` to the multi-one.~~

Given my limited exposure to Tempo’s internals, I’d appreciate any guidance on whether these routes are viable or if there’s a simpler approach to preserve `allConditions`.