@joe-elliott 2025-01-21
> Also, another edge case raises questions: what happens if a float hits MaxInt/MinInt? In some cases, it might cause jumps between MaxInt and MinInt.

We could try to get tricky here. Like if you do `{ span.IntCol > IntMaxAsFloat }` then we just don't do the int comparison. `{ span.IntCol < IntMaxAsFloat }` would just return all values from the fetch layer. But I'm also fine with the easy path of just not attempting the float/int comparison is the float is outside the bounds of the int column. It feels like an acceptable edge case as long as we document it.

> Done! Let me know if this aligns with what you had in mind.

Yup, I think this communicates better to a future reader what's going on. Thanks for the change.

Ok, I was running your branch on Friday to test and we do have one final thing to figure out. This query does not work:

```
{ span.http.status_code = 200.0 }
```

The reason is b/c we handle this special column here:

https://github.com/grafana/tempo/blob/14efba0afa06cdb3fae54d03d3138c585463641a/tempodb/encoding/vparquet4/block_traceql.go#L1969-L1986

All well known and dedicated columns are strings ... except this one unfortunately. To do this correctly we have to scan both the well known column as well as the general float attribute column if the static value being compared against http status code is a float. To do this performantly I think we will need to build a `UnionIterator` that joins two sub iterators. One that scans the well known column and one that scans the float attribute column with the appropriate predicate.