@joe-elliott 2025-02-13
> Yep, I also think so. Roaming around the code base I got an impression that dedicated columns aren't something by default. I feel I'm missing something.

Dedicated columns need to be configured manually. They allow us to move data from the main attribute columns into their own "dedicated" column. The primary thing we use this for is to move very large attributes out of the main columns like sql queries or large json objects. The secondary use is to isolate important columns for querying. The docs have some details on [how we pick which columns](https://grafana.com/docs/tempo/latest/operations/dedicated_columns/) to configure.

So I reran benchmarks and the regression is fairly intense and likely not something we can accept. I am still working on determining what is causing the regression and seeing if we can improve performance, but my heads down time is limited. This is really cool work and it's so close to mergeable, but the `span.http.status_code` column is frequently queried and we can't sustain a hit like the benchmarks are suggesting. I have not given up hope yet. I'm chasing a few threads, but I do wonder if this work will be more easily mergeable once we drop the "well known" columns. I'm starting a discussion about that here: 

https://github.com/grafana/tempo/issues/4694

<details>
<summary>benches</summary>

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/tempo/tempodb/encoding/vparquet4
cpu: Apple M3 Pro
                                                    │ before.txt  │              after.txt              │
                                                    │   sec/op    │   sec/op     vs base                │
BackendBlockTraceQL/spanAttValMatch-11                82.76m ± 1%   81.73m ± 2%   -1.25% (p=0.011 n=10)
BackendBlockTraceQL/spanAttValNoMatch-11              5.468m ± 1%   5.503m ± 1%   +0.64% (p=0.029 n=10)
BackendBlockTraceQL/spanAttIntrinsicMatch-11          58.59m ± 0%   58.41m ± 1%   -0.30% (p=0.005 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-11        5.383m ± 2%   5.383m ± 1%        ~ (p=0.579 n=10)
BackendBlockTraceQL/resourceAttValMatch-11            394.6m ± 1%   392.6m ± 1%        ~ (p=0.052 n=10)
BackendBlockTraceQL/resourceAttValNoMatch-11          5.095m ± 0%   5.119m ± 0%   +0.46% (p=0.009 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-11      24.86m ± 0%   24.71m ± 0%   -0.59% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch#01-11   5.170m ± 0%   5.167m ± 0%        ~ (p=0.579 n=10)
BackendBlockTraceQL/traceOrMatch-11                   240.7m ± 1%   362.6m ± 0%  +50.66% (p=0.000 n=10)
BackendBlockTraceQL/traceOrNoMatch-11                 241.4m ± 0%   364.0m ± 0%  +50.74% (p=0.000 n=10)
BackendBlockTraceQL/mixedValNoMatch-11                178.4m ± 0%   180.6m ± 0%   +1.18% (p=0.000 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd-11          5.009m ± 1%   5.021m ± 1%        ~ (p=0.436 n=10)
BackendBlockTraceQL/mixedValMixedMatchOr-11           148.5m ± 0%   149.5m ± 0%   +0.62% (p=0.000 n=10)
BackendBlockTraceQL/count-11                          328.5m ± 0%   327.4m ± 0%   -0.34% (p=0.000 n=10)
BackendBlockTraceQL/struct-11                         429.4m ± 2%   424.2m ± 1%   -1.20% (p=0.011 n=10)
BackendBlockTraceQL/||-11                             152.5m ± 0%   154.4m ± 0%   +1.24% (p=0.000 n=10)
BackendBlockTraceQL/mixed-11                          25.34m ± 0%   25.23m ± 0%   -0.44% (p=0.000 n=10)
BackendBlockTraceQL/complex-11                        4.975m ± 1%   5.002m ± 1%        ~ (p=0.105 n=10)
BackendBlockTraceQL/select-11                         5.009m ± 1%   5.023m ± 1%        ~ (p=0.912 n=10)
geomean                                               40.72m        42.53m        +4.44%

                                                    │  before.txt  │              after.txt              │
                                                    │     B/s      │     B/s       vs base               │
BackendBlockTraceQL/spanAttValMatch-11                270.6Mi ± 1%   274.0Mi ± 2%  +1.27% (p=0.011 n=10)
BackendBlockTraceQL/spanAttValNoMatch-11              309.0Mi ± 1%   307.0Mi ± 1%  -0.64% (p=0.028 n=10)
BackendBlockTraceQL/spanAttIntrinsicMatch-11          397.6Mi ± 0%   398.8Mi ± 1%  +0.30% (p=0.005 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-11        460.2Mi ± 2%   460.2Mi ± 1%       ~ (p=0.579 n=10)
BackendBlockTraceQL/resourceAttValMatch-11            55.69Mi ± 1%   55.98Mi ± 1%       ~ (p=0.050 n=10)
BackendBlockTraceQL/resourceAttValNoMatch-11          176.5Mi ± 0%   175.7Mi ± 0%  -0.46% (p=0.007 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-11      885.8Mi ± 0%   891.1Mi ± 0%  +0.59% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch#01-11   180.6Mi ± 0%   180.7Mi ± 0%       ~ (p=0.579 n=10)
BackendBlockTraceQL/traceOrMatch-11                   6.948Mi ± 0%   6.876Mi ± 0%  -1.03% (p=0.000 n=10)
BackendBlockTraceQL/traceOrNoMatch-11                 6.924Mi ± 0%   6.852Mi ± 0%  -1.03% (p=0.000 n=10)
BackendBlockTraceQL/mixedValNoMatch-11                11.05Mi ± 0%   10.91Mi ± 0%  -1.21% (p=0.000 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd-11          178.2Mi ± 1%   177.8Mi ± 0%       ~ (p=0.436 n=10)
BackendBlockTraceQL/mixedValMixedMatchOr-11           18.57Mi ± 0%   18.46Mi ± 0%  -0.62% (p=0.000 n=10)
BackendBlockTraceQL/count-11                          66.86Mi ± 0%   67.09Mi ± 0%  +0.34% (p=0.000 n=10)
BackendBlockTraceQL/struct-11                         12.72Mi ± 2%   12.87Mi ± 1%  +1.24% (p=0.014 n=10)
BackendBlockTraceQL/||-11                             144.8Mi ± 0%   143.0Mi ± 0%  -1.22% (p=0.000 n=10)
BackendBlockTraceQL/mixed-11                          843.2Mi ± 0%   846.9Mi ± 0%  +0.45% (p=0.000 n=10)
BackendBlockTraceQL/complex-11                        180.9Mi ± 1%   179.9Mi ± 1%       ~ (p=0.093 n=10)
BackendBlockTraceQL/select-11                         179.7Mi ± 1%   179.2Mi ± 1%       ~ (p=0.869 n=10)
geomean                                               101.9Mi        101.8Mi       -0.14%

                                                    │ before.txt  │               after.txt               │
                                                    │  MB_io/op   │  MB_io/op    vs base                  │
BackendBlockTraceQL/spanAttValMatch-11                 23.48 ± 0%    23.48 ± 0%        ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/spanAttValNoMatch-11               1.772 ± 0%    1.772 ± 0%        ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/spanAttIntrinsicMatch-11           24.43 ± 0%    24.43 ± 0%        ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/spanAttIntrinsicNoMatch-11         2.598 ± 0%    2.598 ± 0%        ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/resourceAttValMatch-11             23.04 ± 0%    23.04 ± 0%        ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/resourceAttValNoMatch-11          943.2m ± 0%   943.2m ± 0%        ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/resourceAttIntrinsicMatch-11       23.09 ± 0%    23.09 ± 0%        ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/resourceAttIntrinsicMatch#01-11   979.0m ± 0%   979.0m ± 0%        ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/traceOrMatch-11                    1.753 ± 0%    2.615 ± 0%  +49.17% (p=0.000 n=10)
BackendBlockTraceQL/traceOrNoMatch-11                  1.753 ± 0%    2.615 ± 0%  +49.17% (p=0.000 n=10)
BackendBlockTraceQL/mixedValNoMatch-11                 2.067 ± 0%    2.067 ± 0%        ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/mixedValMixedMatchAnd-11          936.1m ± 0%   936.1m ± 0%        ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/mixedValMixedMatchOr-11            2.893 ± 0%    2.893 ± 0%        ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/count-11                           23.03 ± 0%    23.03 ± 0%        ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/struct-11                          5.726 ± 0%    5.726 ± 0%        ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/||-11                              23.14 ± 0%    23.14 ± 0%        ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/mixed-11                           22.40 ± 0%    22.40 ± 0%        ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/complex-11                        943.7m ± 0%   943.7m ± 0%        ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/select-11                         943.7m ± 0%   943.7m ± 0%        ~ (p=1.000 n=10) ¹
geomean                                                4.351         4.538        +4.30%
¹ all samples are equal

                                                    │  before.txt   │               after.txt                │
                                                    │     B/op      │      B/op       vs base                │
BackendBlockTraceQL/spanAttValMatch-11                45.28Mi ±  1%    45.01Mi ±  1%        ~ (p=0.353 n=10)
BackendBlockTraceQL/spanAttValNoMatch-11              6.693Mi ±  1%    6.716Mi ±  1%        ~ (p=0.796 n=10)
BackendBlockTraceQL/spanAttIntrinsicMatch-11          39.78Mi ±  1%    39.65Mi ±  1%        ~ (p=0.123 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-11        7.021Mi ±  2%    7.032Mi ±  1%        ~ (p=0.912 n=10)
BackendBlockTraceQL/resourceAttValMatch-11            576.8Mi ±  0%    576.0Mi ±  1%        ~ (p=0.481 n=10)
BackendBlockTraceQL/resourceAttValNoMatch-11          5.308Mi ±  1%    5.362Mi ±  1%        ~ (p=0.063 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-11      10.52Mi ±  1%    10.48Mi ±  1%        ~ (p=0.912 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch#01-11   6.768Mi ±  1%    6.763Mi ±  1%        ~ (p=0.912 n=10)
BackendBlockTraceQL/traceOrMatch-11                   9.428Mi ± 20%   12.960Mi ± 12%  +37.46% (p=0.001 n=10)
BackendBlockTraceQL/traceOrNoMatch-11                 10.02Mi ± 23%    13.47Mi ± 15%  +34.47% (p=0.000 n=10)
BackendBlockTraceQL/mixedValNoMatch-11                8.002Mi ± 11%    8.127Mi ±  8%        ~ (p=0.853 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd-11          5.743Mi ±  1%    5.738Mi ±  1%        ~ (p=0.684 n=10)
BackendBlockTraceQL/mixedValMixedMatchOr-11           7.971Mi ±  8%    8.119Mi ±  8%        ~ (p=0.436 n=10)
BackendBlockTraceQL/count-11                          418.7Mi ±  0%    417.9Mi ±  0%        ~ (p=0.165 n=10)
BackendBlockTraceQL/struct-11                         13.66Mi ± 17%    15.59Mi ± 10%        ~ (p=0.143 n=10)
BackendBlockTraceQL/||-11                             16.74Mi ±  5%    17.05Mi ±  4%        ~ (p=0.912 n=10)
BackendBlockTraceQL/mixed-11                          7.048Mi ±  3%    7.015Mi ±  2%        ~ (p=0.165 n=10)
BackendBlockTraceQL/complex-11                        5.613Mi ±  1%    5.499Mi ±  3%        ~ (p=0.075 n=10)
BackendBlockTraceQL/select-11                         5.554Mi ±  2%    5.531Mi ±  2%        ~ (p=0.631 n=10)
geomean                                               14.61Mi          15.21Mi         +4.11%

                                                    │ before.txt  │             after.txt              │
                                                    │  allocs/op  │  allocs/op   vs base               │
BackendBlockTraceQL/spanAttValMatch-11                503.6k ± 0%   498.5k ± 0%  -1.02% (p=0.000 n=10)
BackendBlockTraceQL/spanAttValNoMatch-11              79.47k ± 0%   79.47k ± 0%       ~ (p=0.179 n=10)
BackendBlockTraceQL/spanAttIntrinsicMatch-11          288.1k ± 0%   287.4k ± 0%  -0.25% (p=0.000 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-11        79.42k ± 0%   79.42k ± 0%       ~ (p=0.155 n=10)
BackendBlockTraceQL/resourceAttValMatch-11            3.450M ± 0%   3.445M ± 0%  -0.16% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttValNoMatch-11          79.47k ± 0%   79.47k ± 0%       ~ (p=0.300 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-11      126.1k ± 0%   125.7k ± 0%  -0.31% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch#01-11   79.46k ± 0%   79.46k ± 0%  -0.00% (p=0.002 n=10)
BackendBlockTraceQL/traceOrMatch-11                   86.53k ± 2%   87.34k ± 2%  +0.93% (p=0.019 n=10)
BackendBlockTraceQL/traceOrNoMatch-11                 86.36k ± 0%   87.04k ± 4%  +0.78% (p=0.000 n=10)
BackendBlockTraceQL/mixedValNoMatch-11                80.29k ± 0%   80.28k ± 0%       ~ (p=0.870 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd-11          79.46k ± 0%   79.46k ± 0%  -0.00% (p=0.010 n=10)
BackendBlockTraceQL/mixedValMixedMatchOr-11           80.29k ± 0%   80.28k ± 0%       ~ (p=0.516 n=10)
BackendBlockTraceQL/count-11                          1.991M ± 0%   1.985M ± 0%  -0.26% (p=0.000 n=10)
BackendBlockTraceQL/struct-11                         93.63k ± 7%   93.42k ± 2%       ~ (p=0.529 n=10)
BackendBlockTraceQL/||-11                             161.9k ± 0%   161.5k ± 0%  -0.27% (p=0.000 n=10)
BackendBlockTraceQL/mixed-11                          87.86k ± 0%   87.83k ± 0%  -0.02% (p=0.000 n=10)
BackendBlockTraceQL/complex-11                        79.82k ± 0%   79.86k ± 0%  +0.04% (p=0.000 n=10)
BackendBlockTraceQL/select-11                         79.70k ± 0%   79.70k ± 0%  -0.00% (p=0.016 n=10)
geomean                                               147.6k        147.5k       -0.04%
```

</details>